Improves Gradle speed
